patch from Ben:
The patch to scene.R is semi-cosmetic: it checks for missing values
in data in rgl.primitive(), ignoring those data and giving a warning.
I put this in because I still have a bug that shows up in the
examples for rgl.window2user(); this patch is a workaround that
lets R CMD check finish without an error, but doesn't fix the
bug.


git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rgl@270 79a1b648-8b68-254f-b20b-c0b50f5137bf